Bus position

Returns bus positions for the given route, with an optional search radius. If no parameters are specified, all bus positions are returned.

bus_position(route = NULL, lat = NULL, lon = NULL, radius = 1000)

Arguments

route
Base bus route, e.g.: 70, 10A.

lat
Center point Latitude.

lon
Center point Longitude.

radius
Meters to include in the search area for lat and lon.

Format

A tibble with 13 variables:

vehicle

Unique identifier. Usually visible on the bus itself.

lat

Last reported Latitude of the bus.

lon

Last reported Longitude of the bus.

dist

Distance in meters from provided coordinates.

deviation

Deviation, in minutes, from schedule.

time

Date and time (EST) of last position update.

trip

Unique trip ID. Used with the schedule-related functions.

route

Base route name as shown on the bus.

direction

General direction of the trip, not the bus itself

head

Destination of the bus.

start

Scheduled start time (EST) of the bus's current trip.

end

Scheduled end time (EST) of the bus's current trip.

block

Source

https://api.wmata.com/Bus.svc/json/jBusPositions

Details

Note that the route parameter accepts only base route names and no variations, i.e.: use 10A instead of 10Av1 or 10Av2.