Bus route and stop methods

Returns schedules for a given route variant for a given date.

bus_schedule(route = NULL, variants = TRUE, date = NULL)

Arguments

route
Bus route variant, e.g.: 70, 10A, 10Av1, etc.

variants
Should variations of route be included?. For example, if B30 is specified and TRUE (default), data for all variations of B30 such as B30v1, B30v2, etc. will be returned.

date
Date for which to retrieve route and stop information.

Format

A tibble with 10 variables:

route_id

The bus route ID

direction

General route direction

trip_head

Destination of the bus.

trip_id

Trip identifier.

start_time

Scheduled start date and time (EST) for this trip.

end_time

Scheduled end date and time (EST) for this trip.

stop_seq

Order of the stop in the sequence of StopTimes.

stop_id

7-digit regional ID. If unavailable, 0 or NULL.

stop_name

Stop name. May be different from what is displayed.

stop_time

Scheduled departure date and time (EST) from this stop.

Source

https://api.wmata.com/Bus.svc/json/jRouteSchedule