Rail Station to Station Information

Returns a distance, fare information, and estimated travel time between any two stations, including those on different lines. Omit both parameters to retrieve data for all stations.

rail_destination(
  FromStationCode = NULL,
  ToStationCode = NULL,
  api_key = wmata_key()
)

Arguments

FromStationCode
Station code for the origin station. Use the rail_stations() function to return a list of all station codes.

ToStationCode
Station code for the destination station. Use the rail_stations() function to return a list of all station codes.

api_key
Subscription key which provides access to this API. Found in your Profile. Defaults Sys.getenv("WMATA_KEY") via wmata_key().

Format

A tibble 1 row per with variables:

SourceStation

Origin station code. Use this value in other rail-related APIs to retrieve data about a station.

DestinationStation

Destination station code. Use this value in other rail-related APIs to retrieve data about a station.

CompositeMiles

Average of distance traveled between two stations and straight-line distance (as used for WMATA fare calculations). For more details, please refer to WMATA's Tariff on Fares.

RailTime

Estimated travel time (schedule time) in minutes between the source and destination station. This is not correlated to minutes (Min) in Real-Time Rail Predictions.

PeakTime

Fare during peak times (weekdays from opening to 9:30 AM and 3-7 PM (EST), and weekends from midnight to closing).

OffPeakTime

Fare during off-peak times (times other than the ones described below).

SeniorDisabled

Reduced fare for senior citizens or people with disabilities.

Value

A data frame containing station to station information

See also

Examples

if (FALSE) {
rail_destination("A01", "A08")
}