Rail Station List

Returns a list of station location and address information based on a given LineCode. Use NULL (default) to return all stations. The response is an data frame identical to that returned in the Station Information method.

rail_stations(LineCode = NULL, api_key = wmata_key())

Format

A tibble 1 row per station with 10 variables:

StationCode

Station code for this station. Use this value in other rail-related APIs to retrieve data about a station.

StationName

Full name for this station, as shown on the WMATA website.

StationTogether

For stations with multiple platforms (e.g.: Gallery Place, Fort Totten, L'Enfant Plaza, and Metro Center), the additional StationCode will be listed here.

LineCodes

Character vector of two-letter abbreviations (e.g.: RD, BL, YL, OR, GR, or SV) served by this station. If the station has an additional platform, the lines served by the other platform are listed in the LineCodes values for the record associated with the StationCode found in StationTogether.

Lat

Latitude.

Lon

Longitude.

Street

Street address (for GPS use).

City

City.

State

State (abbreviated).

Zip

Zip code.

Arguments

LineCode

Two-letter line code abbreviation, or NULL (default):

RD - Red

YL - Yellow

GR - Green

BL - Blue

OR - Orange

SV - Silver

api_key

Subscription key which provides access to this API. Defaults Sys.getenv("WMATA_KEY") via wmata_key().

Value

A data frame of stations on a rail line.

See also

Examples

if (FALSE) {
rail_stations("RD")
}