Track Circuits

Returns a list of all track circuits including those on pocket tracks and crossovers. Each track circuit may include references to its right and left neighbors.

track_circuits()

Format

A nested tibble with 1 row per train and 9 variables:

Track

Track number. 1 and 2 denote "main" lines, while 0 and 3 are connectors (between different types of tracks) and pocket tracks, respectively.

CircuitId

An internal system-wide uniquely identifiable circuit number.

Neighbors

Data frame containing track circuit neighbor information. Note that some track circuits have no neighbors in one direction. All track circuits have at least one neighbor.

NeighborType: Left or Right neighbor group. Generally speaking, left neighbors are to the west and south, while right neighbors are to the east/north.

CircuitIds: Data frame containing neighboring circuit IDs as list column.

Value

A data frame of nested track circuit information.

Details

Please refer to this page for additional details.

See also

Examples

if (FALSE) {
track_circuits()
}