Find a WMATA API key

All calls to the WMATA API must be accompanied by a personal API key. A free key can be obtained by subscribing to the default tier. See browse_wmata_key() for more information on getting a key and storing on your system.

The WMATA also provides a demonstration key. This key should never be used in production applications, it is rate limited and subject to change at any time. The key can be found on this page. If the user has the "rvest" package installed, the wmata_demo() function can be used to try scrape said page and look for the key automatically.

The wmata_validate() function can be used to verify a key is valid.

wmata_key()

wmata_demo()

wmata_validate(key = wmata_key())

browse_wmata_key()

Arguments

key
A WMATA API key to validate, defaults to wamta_key().

Details

Default tier sufficient for most casual developers. Rate limited to 10 calls/second and 50,000 calls per day. This product contains 8 APIs:

Bus Route and Stop Methods

GTFS

Incidents

Misc Methods

Rail Station Information

Real-Time Bus Predictions

Real-Time Rail Predictions

Train Positions